Update AWS region in GitHub workflow
The AWS_DEFAULT_REGION value in the container.yml file no longer pulls from GitHub secrets but has been hard-coded to 'us-east-1'. While this restricts the flexibility of changing regions, it enhances the consistency of the workflow, by ensuring that the code will only ever run in this specific AWS region.